[CI] Split core/core_slow/extended language generation tests into directories
Complete the folder-sharding layout for the remaining generation tests.
Common test_models is split by model group into core/ (core_model, not
slow), core_slow/ (core_model and slow) and the extended/05..08 slices,
each importing the shared, non-collected _common_models.py body with its
exact pytest.param marks and the ROCm use_rocm_aiter axis. Granite splits
into extended/02 (PowerMoE-3b whole) and extended/03 (PowerLM-3b, the two
swash models and the SWA unit tests) over the shared _granite_models.py
body. test_mistral.py, test_gemma.py (incl. the cpu_test KV-sharing unit
tests with the dist_init fixture) and test_phimoe.py (skip marks intact)
move whole into extended/01, 04 and 06. The old test_common.py,
test_granite.py and the moved files are removed so nothing
double-collects; tests/tools/test_language_layout.py now proves all three
shared helpers are never collected. The PPC64LE CPU script's three exact
node IDs point at core/ and core_slow/. The extended/ numeric dirs join
the mypy exclusion (core/ and core_slow/ are valid identifiers and stay
checked). The area YAMLs are unchanged (their test_common.py
source_file_dependencies entries are wired in ticket 08); the existing
marker-based commands select the same case multisets over the new tree.
