+++ tests/models/language/generation/extended/02/test_granite_powermoe.py
@@ -1,0 +1,25 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+
+import pytest
+
+from tests.models.language.generation._granite_models import _test_models
+
+
+@pytest.mark.parametrize("model", ["ibm/PowerMoE-3b"])
+@pytest.mark.parametrize("dtype", ["bfloat16"])
+@pytest.mark.parametrize("max_tokens", [64])
+@pytest.mark.parametrize("num_logprobs", [5])
+@pytest.mark.cpu_model
+def test_models(
+    hf_runner,
+    vllm_runner,
+    example_prompts,
+    model: str,
+    dtype: str,
+    max_tokens: int,
+    num_logprobs: int,
+) -> None:
+    _test_models(
+        hf_runner, vllm_runner, example_prompts, model, dtype, max_tokens, num_logprobs
+    )
