+++ tests/models/language/generation/extended/03/test_granite.py
@@ -1,23 +1,21 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+
 import pytest
 from transformers import GraniteConfig
 
+from tests.models.language.generation._granite_models import _test_models
 from vllm.model_executor.models.granite import granite_layer_attn_params
 
-from ...utils import check_logprobs_close, check_transformers_version
 
-# model -> minimum transformers version, or None if unconstrained
-MODELS = {
-    # TODO(sang): Sliding window should be tested separately.
-    "ibm/PowerLM-3b": None,
-    "ibm/PowerMoE-3b": None,
-    "ibm-granite/granite-swash-2b": "5.15.1",
-    "ibm-granite/granite-swash-3b-a600m": "5.15.1",
-}
-
-
-@pytest.mark.parametrize("model", MODELS)
+@pytest.mark.parametrize(
+    "model",
+    [
+        "ibm/PowerLM-3b",
+        "ibm-granite/granite-swash-2b",
+        "ibm-granite/granite-swash-3b-a600m",
+    ],
+)
 @pytest.mark.parametrize("dtype", ["bfloat16"])
 @pytest.mark.parametrize("max_tokens", [64])
 @pytest.mark.parametrize("num_logprobs", [5])
@@ -31,22 +29,8 @@ def test_models(
     max_tokens: int,
     num_logprobs: int,
 ) -> None:
-    check_transformers_version(model, min_transformers_version=MODELS[model])
-
-    with hf_runner(model, dtype=dtype) as hf_model:
-        hf_outputs = hf_model.generate_greedy_logprobs_limit(
-            example_prompts, max_tokens, num_logprobs
-        )
-
-    with vllm_runner(model, dtype=dtype) as vllm_model:
-        vllm_outputs = vllm_model.generate_greedy_logprobs(
-            example_prompts, max_tokens, num_logprobs
-        )
-    check_logprobs_close(
-        outputs_0_lst=hf_outputs,
-        outputs_1_lst=vllm_outputs,
-        name_0="hf",
-        name_1="vllm",
+    _test_models(
+        hf_runner, vllm_runner, example_prompts, model, dtype, max_tokens, num_logprobs
     )
