+++ tests/models/language/generation/extended/06/test_phimoe.py
@@ -3,10 +3,9 @@
 import pytest
 import torch
 
+from tests.models.utils import check_logprobs_close
+from tests.utils import large_gpu_test
 from vllm.platforms import current_platform
-
-from ....utils import large_gpu_test
-from ...utils import check_logprobs_close
 
 MODELS = [
     "microsoft/Phi-3.5-MoE-instruct",
@@ -62,7 +61,7 @@ def test_phimoe_routing_function():
 
 # There is a known issue that triggers `AttributeError: 'DynamicCache'
 # object has no attribute 'seen_tokens'` when running:
-# `tests/models/language/generation/test_phimoe.py::test_models
+# `tests/models/language/generation/extended/06/test_phimoe.py::test_models
 #   [5-64-bfloat16-microsoft/Phi-3.5-MoE-instruct]`
 # This issue is being investigated and tracked in:
 #   https://huggingface.co/microsoft/Phi-3.5-MoE-instruct/discussions/58
